tls: add support for building against OpenSSL/quictls fork
This adds support for building against the "quictls" fork of OpenSSL
instead of BoringSSL, to provide crypto and TLS support.

Because upstream OpenSSL doesn't (yet?) expose an API for implementing
the QUIC handshake, like BoringSSL does, a fork of it was created called
quictls.

This functionality can be useful for applications that already use
OpenSSL and where adding BoringSSL on top would create conflicts.
