+++ quiche/src/crypto/mod.rs
@@ -24,12 +24,9 @@
 // NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 // SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-use std::mem::MaybeUninit;
-
 use ring::aead;
 use ring::hkdf;
 
-use libc::c_int;
 use libc::c_void;
 
 use crate::Error;
@@ -71,16 +68,8 @@ pub enum Algorithm {
 }
 
 impl Algorithm {
-    fn get_evp_aead(self) -> *const EVP_AEAD {
-        match self {
-            Algorithm::AES128_GCM => unsafe { EVP_aead_aes_128_gcm() },
-            Algorithm::AES256_GCM => unsafe { EVP_aead_aes_256_gcm() },
-            Algorithm::ChaCha20_Poly1305 => unsafe {
-                EVP_aead_chacha20_poly1305()
-            },
-        }
-    }
-
+    // Note: some vendor-specific methods are implemented by each vendor's
+    // submodule (openssl-quictls / boringssl).
     fn get_ring_hp(self) -> &'static aead::quic::Algorithm {
         match self {
             Algorithm::AES128_GCM => &aead::quic::AES_128,
@@ -126,6 +115,12 @@ impl Algorithm {
     }
 }
 
+#[allow(non_camel_case_types)]
+#[repr(transparent)]
+pub struct EVP_AEAD {
+    _unused: c_void,
+}
+
 pub struct Open {
     alg: Algorithm,
 
@@ -137,6 +132,10 @@ pub struct Open {
 }
 
 impl Open {
+    // Note: some vendor-specific methods are implemented by each vendor's
+    // submodule (openssl-quictls / boringssl).
+    pub const DECRYPT: u32 = 0;
+
     pub fn new(
         alg: Algorithm, key: Vec<u8>, iv: Vec<u8>, hp_key: Vec<u8>,
         secret: Vec<u8>,
@@ -146,7 +145,7 @@ impl Open {
 
             header: HeaderProtectionKey::new(alg, hp_key)?,
 
-            packet: PacketKey::new(alg, key, iv)?,
+            packet: PacketKey::new(alg, key, iv, Self::DECRYPT)?,
 
             secret,
         })
@@ -158,50 +157,10 @@ impl Open {
 
             header: HeaderProtectionKey::from_secret(aead, &secret)?,
 
-            packet: PacketKey::from_secret(aead, &secret)?,
+            packet: PacketKey::from_secret(aead, &secret, Self::DECRYPT)?,
 
             secret,
         })
-    }
-
-    pub fn open_with_u64_counter(
-        &self, counter: u64, ad: &[u8], buf: &mut [u8],
-    ) -> Result<usize> {
-        if cfg!(feature = "fuzzing") {
-            return Ok(buf.len());
-        }
-
-        let tag_len = self.alg().tag_len();
-
-        let mut out_len = match buf.len().checked_sub(tag_len) {
-            Some(n) => n,
-            None => return Err(Error::CryptoFail),
-        };
-
-        let max_out_len = out_len;
-
-        let nonce = make_nonce(&self.packet.nonce, counter);
-
-        let rc = unsafe {
-            EVP_AEAD_CTX_open(
-                &self.packet.ctx,   // ctx
-                buf.as_mut_ptr(),   // out
-                &mut out_len,       // out_len
-                max_out_len,        // max_out_len
-                nonce[..].as_ptr(), // nonce
-                nonce.len(),        // nonce_len
-                buf.as_ptr(),       // inp
-                buf.len(),          // in_len
-                ad.as_ptr(),        // ad
-                ad.len(),           // ad_len
-            )
-        };
-
-        if rc != 1 {
-            return Err(Error::CryptoFail);
-        }
-
-        Ok(out_len)
     }
 
     pub fn new_mask(&self, sample: &[u8]) -> Result<[u8; 5]> {
@@ -225,7 +184,8 @@ impl Open {
     pub fn derive_next_packet_key(&self) -> Result<Open> {
         let next_secret = derive_next_secret(self.alg, &self.secret)?;
 
-        let next_packet_key = PacketKey::from_secret(self.alg, &next_secret)?;
+        let next_packet_key =
+            PacketKey::from_secret(self.alg, &next_secret, Self::DECRYPT)?;
 
         Ok(Open {
             alg: self.alg,
@@ -253,6 +213,10 @@ pub struct Seal {
 }
 
 impl Seal {
+    // Note: some vendor-specific methods are implemented by each vendor's
+    // submodule (openssl-quictls / boringssl).
+    const ENCRYPT: u32 = 1;
+
     pub fn new(
         alg: Algorithm, key: Vec<u8>, iv: Vec<u8>, hp_key: Vec<u8>,
         secret: Vec<u8>,
@@ -262,7 +226,7 @@ impl Seal {
 
             header: HeaderProtectionKey::new(alg, hp_key)?,
 
-            packet: PacketKey::new(alg, key, iv)?,
+            packet: PacketKey::new(alg, key, iv, Self::ENCRYPT)?,
 
             secret,
         })
@@ -274,65 +238,10 @@ impl Seal {
 
             header: HeaderProtectionKey::from_secret(aead, &secret)?,
 
-            packet: PacketKey::from_secret(aead, &secret)?,
+            packet: PacketKey::from_secret(aead, &secret, Self::ENCRYPT)?,
 
             secret,
         })
-    }
-
-    pub fn seal_with_u64_counter(
-        &self, counter: u64, ad: &[u8], buf: &mut [u8], in_len: usize,
-        extra_in: Option<&[u8]>,
-    ) -> Result<usize> {
-        if cfg!(feature = "fuzzing") {
-            if let Some(extra) = extra_in {
-                buf[in_len..in_len + extra.len()].copy_from_slice(extra);
-                return Ok(in_len + extra.len());
-            }
-
-            return Ok(in_len);
-        }
-
-        let tag_len = self.alg().tag_len();
-
-        let mut out_tag_len = tag_len;
-
-        let (extra_in_ptr, extra_in_len) = match extra_in {
-            Some(v) => (v.as_ptr(), v.len()),
-
-            None => (std::ptr::null(), 0),
-        };
-
-        // Make sure all the outputs combined fit in the buffer.
-        if in_len + tag_len + extra_in_len > buf.len() {
-            return Err(Error::CryptoFail);
-        }
-
-        let nonce = make_nonce(&self.packet.nonce, counter);
-
-        let rc = unsafe {
-            EVP_AEAD_CTX_seal_scatter(
-                &self.packet.ctx,           // ctx
-                buf.as_mut_ptr(),           // out
-                buf[in_len..].as_mut_ptr(), // out_tag
-                &mut out_tag_len,           // out_tag_len
-                tag_len + extra_in_len,     // max_out_tag_len
-                nonce[..].as_ptr(),         // nonce
-                nonce.len(),                // nonce_len
-                buf.as_ptr(),               // inp
-                in_len,                     // in_len
-                extra_in_ptr,               // extra_in
-                extra_in_len,               // extra_in_len
-                ad.as_ptr(),                // ad
-                ad.len(),                   // ad_len
-            )
-        };
-
-        if rc != 1 {
-            return Err(Error::CryptoFail);
-        }
-
-        Ok(in_len + out_tag_len)
     }
 
     pub fn new_mask(&self, sample: &[u8]) -> Result<[u8; 5]> {
@@ -356,7 +265,8 @@ impl Seal {
     pub fn derive_next_packet_key(&self) -> Result<Seal> {
         let next_secret = derive_next_secret(self.alg, &self.secret)?;
 
-        let next_packet_key = PacketKey::from_secret(self.alg, &next_secret)?;
+        let next_packet_key =
+            PacketKey::from_secret(self.alg, &next_secret, Self::ENCRYPT)?;
 
         Ok(Seal {
             alg: self.alg,
@@ -394,35 +304,6 @@ impl HeaderProtectionKey {
         derive_hdr_key(aead, secret, &mut hp_key)?;
 
         Self::new(aead, hp_key)
-    }
-}
-
-pub struct PacketKey {
-    ctx: EVP_AEAD_CTX,
-
-    nonce: Vec<u8>,
-}
-
-impl PacketKey {
-    pub fn new(alg: Algorithm, key: Vec<u8>, iv: Vec<u8>) -> Result<Self> {
-        Ok(Self {
-            ctx: make_aead_ctx(alg, &key)?,
-
-            nonce: iv,
-        })
-    }
-
-    pub fn from_secret(aead: Algorithm, secret: &[u8]) -> Result<Self> {
-        let key_len = aead.key_len();
-        let nonce_len = aead.nonce_len();
-
-        let mut key = vec![0; key_len];
-        let mut iv = vec![0; nonce_len];
-
-        derive_pkt_key(aead, secret, &mut key)?;
-        derive_pkt_iv(aead, secret, &mut iv)?;
-
-        Self::new(aead, key, iv)
     }
 }
 
@@ -555,7 +436,7 @@ pub fn derive_pkt_key(
 ) -> Result<()> {
     const LABEL: &[u8] = b"quic key";
 
-    let key_len = aead.key_len();
+    let key_len: usize = aead.key_len();
 
     if key_len > out.len() {
         return Err(Error::CryptoFail);
@@ -578,31 +459,6 @@ pub fn derive_pkt_iv(
 
     let secret = hkdf::Prk::new_less_safe(aead.get_ring_digest(), secret);
     hkdf_expand_label(&secret, LABEL, &mut out[..nonce_len])
-}
-
-fn make_aead_ctx(alg: Algorithm, key: &[u8]) -> Result<EVP_AEAD_CTX> {
-    let mut ctx = MaybeUninit::uninit();
-
-    let ctx = unsafe {
-        let aead = alg.get_evp_aead();
-
-        let rc = EVP_AEAD_CTX_init(
-            ctx.as_mut_ptr(),
-            aead,
-            key.as_ptr(),
-            alg.key_len(),
-            alg.tag_len(),
-            std::ptr::null_mut(),
-        );
-
-        if rc != 1 {
-            return Err(Error::CryptoFail);
-        }
-
-        ctx.assume_init()
-    };
-
-    Ok(ctx)
 }
 
 fn hkdf_expand_label(
@@ -645,51 +501,6 @@ impl hkdf::KeyType for ArbitraryOutputLen {
     fn len(&self) -> usize {
         self.0
     }
-}
-
-#[allow(non_camel_case_types)]
-#[repr(transparent)]
-struct EVP_AEAD {
-    _unused: c_void,
-}
-
-// NOTE: This structure is copied from <openssl/aead.h> in order to be able to
-// statically allocate it. While it is not often modified upstream, it needs to
-// be kept in sync.
-#[repr(C)]
-struct EVP_AEAD_CTX {
-    aead: libc::uintptr_t,
-    opaque: [u8; 580],
-    alignment: u64,
-    tag_len: u8,
-}
-
-extern {
-    // EVP_AEAD
-    fn EVP_aead_aes_128_gcm() -> *const EVP_AEAD;
-
-    fn EVP_aead_aes_256_gcm() -> *const EVP_AEAD;
-
-    fn EVP_aead_chacha20_poly1305() -> *const EVP_AEAD;
-
-    // EVP_AEAD_CTX
-    fn EVP_AEAD_CTX_init(
-        ctx: *mut EVP_AEAD_CTX, aead: *const EVP_AEAD, key: *const u8,
-        key_len: usize, tag_len: usize, engine: *mut c_void,
-    ) -> c_int;
-
-    fn EVP_AEAD_CTX_open(
-        ctx: *const EVP_AEAD_CTX, out: *mut u8, out_len: *mut usize,
-        max_out_len: usize, nonce: *const u8, nonce_len: usize, inp: *const u8,
-        in_len: usize, ad: *const u8, ad_len: usize,
-    ) -> c_int;
-
-    fn EVP_AEAD_CTX_seal_scatter(
-        ctx: *const EVP_AEAD_CTX, out: *mut u8, out_tag: *mut u8,
-        out_tag_len: *mut usize, max_out_tag_len: usize, nonce: *const u8,
-        nonce_len: usize, inp: *const u8, in_len: usize, extra_in: *const u8,
-        extra_in_len: usize, ad: *const u8, ad_len: usize,
-    ) -> c_int;
 }
 
 #[cfg(test)]
@@ -813,3 +624,13 @@ mod tests {
         assert_eq!(&hdr_key, &expected_hdr_key);
     }
 }
+
+#[cfg(not(feature = "openssl"))]
+mod boringssl;
+#[cfg(not(feature = "openssl"))]
+use boringssl::*;
+
+#[cfg(feature = "openssl")]
+mod openssl_quictls;
+#[cfg(feature = "openssl")]
+use openssl_quictls::*;
