+++ quiche/src/tls/mod.rs
@@ -34,7 +34,6 @@ use once_cell::sync::Lazy;
 
 use libc::c_char;
 use libc::c_int;
-use libc::c_long;
 use libc::c_uint;
 use libc::c_void;
 
@@ -113,51 +112,6 @@ struct STACK_OF {
     _unused: c_void,
 }
 
-#[allow(non_camel_case_types)]
-#[repr(transparent)]
-struct CRYPTO_BUFFER {
-    _unused: c_void,
-}
-
-#[repr(C)]
-#[allow(non_camel_case_types)]
-struct SSL_QUIC_METHOD {
-    set_read_secret: Option<
-        unsafe extern fn(
-            ssl: *mut SSL,
-            level: crypto::Level,
-            cipher: *const SSL_CIPHER,
-            secret: *const u8,
-            secret_len: usize,
-        ) -> c_int,
-    >,
-
-    set_write_secret: Option<
-        unsafe extern fn(
-            ssl: *mut SSL,
-            level: crypto::Level,
-            cipher: *const SSL_CIPHER,
-            secret: *const u8,
-            secret_len: usize,
-        ) -> c_int,
-    >,
-
-    add_handshake_data: Option<
-        unsafe extern fn(
-            ssl: *mut SSL,
-            level: crypto::Level,
-            data: *const u8,
-            len: usize,
-        ) -> c_int,
-    >,
-
-    flush_flight: Option<extern fn(ssl: *mut SSL) -> c_int>,
-
-    send_alert: Option<
-        extern fn(ssl: *mut SSL, level: crypto::Level, alert: u8) -> c_int,
-    >,
-}
-
 #[cfg(test)]
 #[repr(C)]
 #[allow(non_camel_case_types)]
@@ -168,43 +122,6 @@ enum ssl_private_key_result_t {
     ssl_private_key_failure,
 }
 
-#[cfg(test)]
-#[repr(C)]
-#[allow(non_camel_case_types)]
-struct SSL_PRIVATE_KEY_METHOD {
-    sign: Option<
-        unsafe extern fn(
-            ssl: *mut SSL,
-            out: *mut u8,
-            out_len: *mut usize,
-            max_out: usize,
-            signature_algorithm: u16,
-            r#in: *const u8,
-            in_len: usize,
-        ) -> ssl_private_key_result_t,
-    >,
-
-    decrypt: Option<
-        unsafe extern fn(
-            ssl: *mut SSL,
-            out: *mut u8,
-            out_len: *mut usize,
-            max_out: usize,
-            r#in: *const u8,
-            in_len: usize,
-        ) -> ssl_private_key_result_t,
-    >,
-
-    complete: Option<
-        unsafe extern fn(
-            ssl: *mut SSL,
-            out: *mut u8,
-            out_len: *mut usize,
-            max_out: usize,
-        ) -> ssl_private_key_result_t,
-    >,
-}
-
 /// BoringSSL ex_data index for quiche connections.
 ///
 /// TODO: replace with `std::sync::LazyLock` when stable.
@@ -212,17 +129,11 @@ pub static QUICHE_EX_DATA_INDEX: Lazy<c_int> = Lazy::new(|| unsafe {
     SSL_get_ex_new_index(0, ptr::null(), ptr::null(), ptr::null(), ptr::null())
 });
 
-static QUICHE_STREAM_METHOD: SSL_QUIC_METHOD = SSL_QUIC_METHOD {
-    set_read_secret: Some(set_read_secret),
-    set_write_secret: Some(set_write_secret),
-    add_handshake_data: Some(add_handshake_data),
-    flush_flight: Some(flush_flight),
-    send_alert: Some(send_alert),
-};
-
 pub struct Context(*mut SSL_CTX);
 
 impl Context {
+    // Note: some vendor-specific methods are implemented by each vendor's
+    // submodule (openssl-quictls / boringssl).
     pub fn new() -> Result<Context> {
         unsafe {
             let ctx_raw = SSL_CTX_new(TLS_method());
@@ -366,6 +277,8 @@ impl Context {
         // false -> 0x00 SSL_VERIFY_NONE
         let mode = i32::from(verify);
 
+        // Note: Base on two used modes(see above), it seems ok for both, bssl and
+        // ossl. If mode needs to be ored then it may need to be adjusted.
         unsafe {
             SSL_CTX_set_verify(self.as_mut_ptr(), mode, None);
         }
@@ -414,14 +327,6 @@ impl Context {
         })
     }
 
-    pub fn set_early_data_enabled(&mut self, enabled: bool) {
-        let enabled = i32::from(enabled);
-
-        unsafe {
-            SSL_CTX_set_early_data_enabled(self.as_mut_ptr(), enabled);
-        }
-    }
-
     fn as_mut_ptr(&mut self) -> *mut SSL_CTX {
         self.0
     }
@@ -450,6 +355,8 @@ pub struct Handshake {
 }
 
 impl Handshake {
+    // Note: some vendor-specific methods are implemented by each vendor's
+    // submodule (openssl-quictls / boringssl).
     #[cfg(feature = "ffi")]
     pub unsafe fn from_ptr(ssl: *mut c_void) -> Handshake {
         Handshake::new(ssl as *mut SSL)
@@ -512,16 +419,6 @@ impl Handshake {
     pub fn set_quic_method(&mut self) -> Result<()> {
         map_result(unsafe {
             SSL_set_quic_method(self.as_mut_ptr(), &QUICHE_STREAM_METHOD)
-        })
-    }
-
-    pub fn set_quic_early_data_context(&mut self, context: &[u8]) -> Result<()> {
-        map_result(unsafe {
-            SSL_set_quic_early_data_context(
-                self.as_mut_ptr(),
-                context.as_ptr(),
-                context.len(),
-            )
         })
     }
 
@@ -612,28 +509,6 @@ impl Handshake {
         s.to_str().ok()
     }
 
-    pub fn set_session(&mut self, session: &[u8]) -> Result<()> {
-        unsafe {
-            let ctx = SSL_get_SSL_CTX(self.as_ptr());
-
-            if ctx.is_null() {
-                return Err(Error::TlsFail);
-            }
-
-            let session =
-                SSL_SESSION_from_bytes(session.as_ptr(), session.len(), ctx);
-
-            if session.is_null() {
-                return Err(Error::TlsFail);
-            }
-
-            let rc = SSL_set_session(self.as_mut_ptr(), session);
-            SSL_SESSION_free(session);
-
-            map_result(rc)
-        }
-    }
-
     pub fn provide_data(
         &mut self, level: crypto::Level, buf: &[u8],
     ) -> Result<()> {
@@ -674,10 +549,6 @@ impl Handshake {
         self.map_result_ssl(rc)
     }
 
-    pub fn reset_early_data_reject(&mut self) {
-        unsafe { SSL_reset_early_data_reject(self.as_mut_ptr()) };
-    }
-
     pub fn write_level(&self) -> crypto::Level {
         unsafe { SSL_quic_write_level(self.as_ptr()) }
     }
@@ -689,141 +560,10 @@ impl Handshake {
         get_cipher_from_ptr(cipher.ok()?).ok()
     }
 
-    pub fn curve(&self) -> Option<String> {
-        let curve = unsafe {
-            let curve_id = SSL_get_curve_id(self.as_ptr());
-            if curve_id == 0 {
-                return None;
-            }
-
-            let curve_name = SSL_get_curve_name(curve_id);
-            match ffi::CStr::from_ptr(curve_name).to_str() {
-                Ok(v) => v,
-
-                Err(_) => return None,
-            }
-        };
-
-        Some(curve.to_string())
-    }
-
-    pub fn sigalg(&self) -> Option<String> {
-        let sigalg = unsafe {
-            let sigalg_id = SSL_get_peer_signature_algorithm(self.as_ptr());
-            if sigalg_id == 0 {
-                return None;
-            }
-
-            let sigalg_name = SSL_get_signature_algorithm_name(sigalg_id, 1);
-            match ffi::CStr::from_ptr(sigalg_name).to_str() {
-                Ok(v) => v,
-
-                Err(_) => return None,
-            }
-        };
-
-        Some(sigalg.to_string())
-    }
-
-    pub fn peer_cert_chain(&self) -> Option<Vec<&[u8]>> {
-        let cert_chain = unsafe {
-            let chain =
-                map_result_ptr(SSL_get0_peer_certificates(self.as_ptr())).ok()?;
-
-            let num = sk_num(chain);
-            if num == 0 {
-                return None;
-            }
-
-            let mut cert_chain = vec![];
-            for i in 0..num {
-                let buffer =
-                    map_result_ptr(sk_value(chain, i) as *const CRYPTO_BUFFER)
-                        .ok()?;
-
-                let out_len = CRYPTO_BUFFER_len(buffer);
-                if out_len == 0 {
-                    return None;
-                }
-
-                let out = CRYPTO_BUFFER_data(buffer);
-                let slice = slice::from_raw_parts(out, out_len);
-
-                cert_chain.push(slice);
-            }
-
-            cert_chain
-        };
-
-        Some(cert_chain)
-    }
-
-    pub fn peer_cert(&self) -> Option<&[u8]> {
-        let peer_cert = unsafe {
-            let chain =
-                map_result_ptr(SSL_get0_peer_certificates(self.as_ptr())).ok()?;
-            if sk_num(chain) == 0 {
-                return None;
-            }
-
-            let buffer =
-                map_result_ptr(sk_value(chain, 0) as *const CRYPTO_BUFFER)
-                    .ok()?;
-
-            let out_len = CRYPTO_BUFFER_len(buffer);
-            if out_len == 0 {
-                return None;
-            }
-
-            let out = CRYPTO_BUFFER_data(buffer);
-            slice::from_raw_parts(out, out_len)
-        };
-
-        Some(peer_cert)
-    }
-
     #[cfg(test)]
     pub fn set_options(&mut self, opts: u32) {
         unsafe {
             SSL_set_options(self.as_mut_ptr(), opts);
-        }
-    }
-
-    // Only used for testing handling of failure during key signing.
-    #[cfg(test)]
-    pub fn set_failing_private_key_method(&mut self) {
-        extern fn failing_sign(
-            _ssl: *mut SSL, _out: *mut u8, _out_len: *mut usize, _max_out: usize,
-            _signature_algorithm: u16, _in: *const u8, _in_len: usize,
-        ) -> ssl_private_key_result_t {
-            ssl_private_key_result_t::ssl_private_key_failure
-        }
-
-        extern fn failing_decrypt(
-            _ssl: *mut SSL, _out: *mut u8, _out_len: *mut usize, _max_out: usize,
-            _in: *const u8, _in_len: usize,
-        ) -> ssl_private_key_result_t {
-            ssl_private_key_result_t::ssl_private_key_failure
-        }
-
-        extern fn failing_complete(
-            _ssl: *mut SSL, _out: *mut u8, _out_len: *mut usize, _max_out: usize,
-        ) -> ssl_private_key_result_t {
-            ssl_private_key_result_t::ssl_private_key_failure
-        }
-
-        static QUICHE_PRIVATE_KEY_METHOD: SSL_PRIVATE_KEY_METHOD =
-            SSL_PRIVATE_KEY_METHOD {
-                decrypt: Some(failing_decrypt),
-                sign: Some(failing_sign),
-                complete: Some(failing_complete),
-            };
-
-        unsafe {
-            SSL_set_private_key_method(
-                self.as_mut_ptr(),
-                &QUICHE_PRIVATE_KEY_METHOD,
-            );
         }
     }
 
@@ -833,10 +573,6 @@ impl Handshake {
 
     pub fn is_resumed(&self) -> bool {
         unsafe { SSL_session_reused(self.as_ptr()) == 1 }
-    }
-
-    pub fn is_in_early_data(&self) -> bool {
-        unsafe { SSL_in_early_data(self.as_ptr()) == 1 }
     }
 
     pub fn clear(&mut self) -> Result<()> {
@@ -1168,15 +904,26 @@ extern fn select_alpn(
     ssl: *mut SSL, out: *mut *const u8, out_len: *mut u8, inp: *mut u8,
     in_len: c_uint, _arg: *mut c_void,
 ) -> c_int {
+    // SSL_TLSEXT_ERR_OK 0
+    // SSL_TLSEXT_ERR_ALERT_WARNING 1
+    // SSL_TLSEXT_ERR_ALERT_FATAL 2
+    // SSL_TLSEXT_ERR_NOACK 3
+
+    // Boringssl internally overwrite the return value from this callback, if the
+    // returned value is SSL_TLSEXT_ERR_NOACK and is quic, then the value gets
+    // overwritten to SSL_TLSEXT_ERR_ALERT_FATAL. In contrast openssl/quictls does
+    // not do that, so we need to explicitly respond with
+    // SSL_TLSEXT_ERR_ALERT_FATAL in case it is needed.
+    // TLS_ERROR is redefined for each vendor.
     let ex_data = match get_ex_data_from_ptr::<ExData>(ssl, *QUICHE_EX_DATA_INDEX)
     {
         Some(v) => v,
 
-        None => return 3, // SSL_TLSEXT_ERR_NOACK
+        None => return TLS_ERROR,
     };
 
     if ex_data.application_protos.is_empty() {
-        return 3; // SSL_TLSEXT_ERR_NOACK
+        return TLS_ERROR;
     }
 
     let mut protos = octets::Octets::with_slice(unsafe {
@@ -1210,7 +957,7 @@ extern fn select_alpn(
         }
     }
 
-    3 // SSL_TLSEXT_ERR_NOACK
+    TLS_ERROR
 }
 
 extern fn new_session(ssl: *mut SSL, session: *mut SSL_SESSION) -> c_int {
@@ -1225,18 +972,9 @@ extern fn new_session(ssl: *mut SSL, session: *mut SSL_SESSION) -> c_int {
     let peer_params = handshake.quic_transport_params();
 
     // Serialize session object into buffer.
-    let session_bytes = unsafe {
-        let mut out: *mut u8 = std::ptr::null_mut();
-        let mut out_len: usize = 0;
-
-        if SSL_SESSION_to_bytes(session, &mut out, &mut out_len) == 0 {
-            return 0;
-        }
-
-        let session_bytes = std::slice::from_raw_parts(out, out_len).to_vec();
-        OPENSSL_free(out as *mut c_void);
-
-        session_bytes
+    let session_bytes = match get_session_bytes(session) {
+        Ok(v) => v,
+        Err(_) => return 0,
     };
 
     let mut buffer =
@@ -1274,21 +1012,21 @@ extern fn new_session(ssl: *mut SSL, session: *mut SSL_SESSION) -> c_int {
     0
 }
 
-fn map_result(bssl_result: c_int) -> Result<()> {
+pub fn map_result(bssl_result: c_int) -> Result<()> {
     match bssl_result {
         1 => Ok(()),
         _ => Err(Error::TlsFail),
     }
 }
 
-fn map_result_zero_is_success(bssl_result: c_int) -> Result<()> {
+pub fn map_result_zero_is_success(bssl_result: c_int) -> Result<()> {
     match bssl_result {
         0 => Ok(()),
         _ => Err(Error::TlsFail),
     }
 }
 
-fn map_result_ptr<'a, T>(bssl_result: *const T) -> Result<&'a T> {
+pub fn map_result_ptr<'a, T>(bssl_result: *const T) -> Result<&'a T> {
     match unsafe { bssl_result.as_ref() } {
         Some(v) => Ok(v),
         None => Err(Error::TlsFail),
@@ -1307,6 +1045,9 @@ fn log_ssl_error() {
 }
 
 extern {
+    // Note: some vendor-specific methods are implemented by each vendor's
+    // submodule (openssl-quictls / boringssl).
+
     // SSL_METHOD
     fn TLS_method() -> *const SSL_METHOD;
 
@@ -1344,10 +1085,6 @@ extern {
         cb: Option<unsafe extern fn(ssl: *const SSL, line: *const c_char)>,
     );
 
-    fn SSL_CTX_set_tlsext_ticket_keys(
-        ctx: *mut SSL_CTX, key: *const u8, key_len: usize,
-    ) -> c_int;
-
     fn SSL_CTX_set_alpn_protos(
         ctx: *mut SSL_CTX, protos: *const u8, protos_len: usize,
     ) -> c_int;
@@ -1367,22 +1104,12 @@ extern {
         arg: *mut c_void,
     );
 
-    fn SSL_CTX_set_early_data_enabled(ctx: *mut SSL_CTX, enabled: i32);
-
-    fn SSL_CTX_set_session_cache_mode(ctx: *mut SSL_CTX, mode: c_int) -> c_int;
-
     fn SSL_CTX_sess_set_new_cb(
         ctx: *mut SSL_CTX,
         cb: Option<
             unsafe extern fn(ssl: *mut SSL, session: *mut SSL_SESSION) -> c_int,
         >,
     );
-
-    // SSL
-    fn SSL_get_ex_new_index(
-        argl: c_long, argp: *const c_void, unused: *const c_void,
-        dup_unused: *const c_void, free_func: *const c_void,
-    ) -> c_int;
 
     fn SSL_new(ctx: *mut SSL_CTX) -> *mut SSL;
 
@@ -1398,26 +1125,11 @@ extern {
 
     fn SSL_get_current_cipher(ssl: *const SSL) -> *const SSL_CIPHER;
 
-    fn SSL_get_curve_id(ssl: *const SSL) -> u16;
-    fn SSL_get_curve_name(curve: u16) -> *const c_char;
-
-    fn SSL_get_peer_signature_algorithm(ssl: *const SSL) -> u16;
-    fn SSL_get_signature_algorithm_name(
-        sigalg: u16, include_curve: i32,
-    ) -> *const c_char;
-
     fn SSL_set_session(ssl: *mut SSL, session: *mut SSL_SESSION) -> c_int;
 
     fn SSL_get_SSL_CTX(ssl: *const SSL) -> *mut SSL_CTX;
 
-    fn SSL_get0_peer_certificates(ssl: *const SSL) -> *const STACK_OF;
-
-    fn SSL_set_min_proto_version(ssl: *mut SSL, version: u16) -> c_int;
-    fn SSL_set_max_proto_version(ssl: *mut SSL, version: u16) -> c_int;
-
     fn SSL_set_quiet_shutdown(ssl: *mut SSL, mode: c_int);
-
-    fn SSL_set_tlsext_host_name(ssl: *mut SSL, name: *const c_char) -> c_int;
 
     fn SSL_set_quic_transport_params(
         ssl: *mut SSL, params: *const u8, params_len: usize,
@@ -1429,17 +1141,8 @@ extern {
 
     fn SSL_set_quic_use_legacy_codepoint(ssl: *mut SSL, use_legacy: c_int);
 
-    fn SSL_set_quic_early_data_context(
-        ssl: *mut SSL, context: *const u8, context_len: usize,
-    ) -> c_int;
-
     #[cfg(test)]
     fn SSL_set_options(ssl: *mut SSL, opts: u32) -> u32;
-
-    #[cfg(test)]
-    fn SSL_set_private_key_method(
-        ssl: *mut SSL, key_method: *const SSL_PRIVATE_KEY_METHOD,
-    );
 
     fn SSL_get_peer_quic_transport_params(
         ssl: *const SSL, out_params: *mut *const u8, out_params_len: *mut usize,
@@ -1457,8 +1160,6 @@ extern {
 
     fn SSL_process_quic_post_handshake(ssl: *mut SSL) -> c_int;
 
-    fn SSL_reset_early_data_reject(ssl: *mut SSL);
-
     fn SSL_do_handshake(ssl: *mut SSL) -> c_int;
 
     fn SSL_quic_write_level(ssl: *const SSL) -> crypto::Level;
@@ -1466,8 +1167,6 @@ extern {
     fn SSL_session_reused(ssl: *const SSL) -> c_int;
 
     fn SSL_in_init(ssl: *const SSL) -> c_int;
-
-    fn SSL_in_early_data(ssl: *const SSL) -> c_int;
 
     fn SSL_clear(ssl: *mut SSL) -> c_int;
 
@@ -1477,13 +1176,6 @@ extern {
     fn SSL_CIPHER_get_id(cipher: *const SSL_CIPHER) -> c_uint;
 
     // SSL_SESSION
-    fn SSL_SESSION_to_bytes(
-        session: *const SSL_SESSION, out: *mut *mut u8, out_len: *mut usize,
-    ) -> c_int;
-
-    fn SSL_SESSION_from_bytes(
-        input: *const u8, input_len: usize, ctx: *const SSL_CTX,
-    ) -> *mut SSL_SESSION;
 
     fn SSL_SESSION_free(session: *mut SSL_SESSION);
 
@@ -1502,19 +1194,23 @@ extern {
     #[cfg(windows)]
     fn d2i_X509(px: *mut X509, input: *const *const u8, len: c_int) -> *mut X509;
 
-    // STACK_OF
-    fn sk_num(stack: *const STACK_OF) -> usize;
-    fn sk_value(stack: *const STACK_OF, idx: usize) -> *mut c_void;
-
-    // CRYPTO_BUFFER
-    fn CRYPTO_BUFFER_len(buffer: *const CRYPTO_BUFFER) -> usize;
-    fn CRYPTO_BUFFER_data(buffer: *const CRYPTO_BUFFER) -> *const u8;
-
     // ERR
     fn ERR_peek_error() -> c_uint;
 
     fn ERR_error_string_n(err: c_uint, buf: *mut c_char, len: usize);
 
     // OPENSSL
+    #[allow(dead_code)]
     fn OPENSSL_free(ptr: *mut c_void);
+
 }
+
+#[cfg(not(feature = "openssl"))]
+mod boringssl;
+#[cfg(not(feature = "openssl"))]
+use boringssl::*;
+
+#[cfg(feature = "openssl")]
+mod openssl_quictls;
+#[cfg(feature = "openssl")]
+use openssl_quictls::*;
